audit_docker_mvp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.10.0
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.4
bitarray==3.6.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cytoolz==1.0.1
edge-tts==7.0.2
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.116.1
fastapi-cache2[redis]==0.2.2
fastapi-limiter==0.1.6
frozenlist==1.6.0
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
hexbytes==1.3.1
httplib2==0.22.0
idna==3.10
multidict==6.4.4
oauthlib==3.2.2
outcome==1.3.0.post0
parsimonious==0.10.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
pyunormalize==16.0.0
redis==4.6.0
regex==2025.7.34
requests==2.32.4
requests-oauthlib==2.0.0
rlp==4.1.0
rsa==4.9.1
selenium==4.33.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
srt==3.5.3
starlette==0.47.2
tabulate==0.9.0
toolz==1.0.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.23.2
web3==7.13.0
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0